Move komodo_importpubkey
