+++ src/komodo_validation015.h
@@ -97,19 +97,6 @@ int32_t gettxout_scriptPubKey(int32_t height,uint8_t *scriptPubKey,int32_t maxsi
     return(-1);
 }
 
-#include <wallet/wallet.h>
-int32_t komodo_importpubkey(std::string pubkeyspend)
-{
-    CWallet * const pwallet = vpwallets[0];
-    std::string strLabel = pubkeyspend;
-    if ( pwallet != 0 )
-    {
-        LOCK2(cs_main, pwallet->cs_wallet);
-        std::vector<unsigned char> data(ParseHex(pubkeyspend));
-        pwallet->AddWatchOnly(CScript(data.begin(), data.end()));
-    }
-}
-
 // following is ported from libtom
 /* LibTomCrypt, modular cryptographic library -- Tom St Denis
  *
