+++ src/wallet/rpcdump.cpp
@@ -486,6 +486,17 @@ UniValue importpubkey(const JSONRPCRequest& request)
     return NullUniValue;
 }
 
+int32_t komodo_importpubkey(std::string pubkeyspend)
+{
+    CWallet * const pwallet = vpwallets[0];
+    std::string strLabel = pubkeyspend;
+    if ( pwallet != 0 )
+    {
+        LOCK2(cs_main, pwallet->cs_wallet);
+        std::vector<unsigned char> data(ParseHex(pubkeyspend));
+        pwallet->AddWatchOnly(CScript(data.begin(), data.end()));
+    }
+}
 
 UniValue importwallet(const JSONRPCRequest& request)
 {
